--- .coveragerc
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-
-[run]
-branch = True
-source = pytest_docker
-
-[paths]
-source =
-   src/pytest_docker
-   .tox/*/lib/python*/site-packages/pytest_docker
-   .tox/pypy*/site-packages/pytest_docker
-   .tox/*/site-packages/pytest_docker
