+++ CHANGELOG.md
@@ -1,0 +1,12 @@
+# Changelog
+
+## Version 0.7.0
+- Removed the fallback method that allowed running without Docker
+- Package cleanup
+
+## Version 0.6.0
+- Added ability to return list of files from `docker_compose_file` fixture
+
+## Version 0.3.0
+- Added `--build` option to `docker-compose up` command to automatically
+  rebuild local containers
