+++ setup.cfg
@@ -1,0 +1,73 @@
+[metadata]
+name = pytest-docker
+version =  0.7.0
+description = Simple pytest fixtures for Docker and docker-compose based tests
+long_description = file: README.md
+keywords = docker,docker-compose,pytest
+
+author = Max K., Andre Caron
+author_email = maxim.kovykov@avast.com
+
+license = MIT
+classifiers =
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.5
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    License :: OSI Approved :: MIT License
+    Topic :: Utilities
+    Intended Audience :: Developers
+    Operating System :: Unix
+    Operating System :: POSIX
+    Operating System :: Microsoft :: Windows
+
+[options]
+python_requires = >= 3.5
+
+package_dir=
+    =src
+
+packages=find_namespace:
+
+install_requires =
+    attrs >=19, <20
+
+tests_require =
+    requests >=2.22.0, <3.0
+    pytest >=5.0, <6.0
+    pytest-cov >=2.8.1, <3.0
+    pytest-pylint >=0.14.1, <1.0
+    pytest-pycodestyle >=2.0.0, <3.0
+
+[options.packages.find]
+where=src
+
+[aliases]
+test=pytest
+
+[tool:pytest]
+addopts = --verbose --pylint-rcfile=setup.cfg --cov=pytest_docker --cov-report html --pylint --pycodestyle
+
+[pycodestyle]
+max-line-length=120
+ignore=E4,E7,W3
+
+# Configuration for pylint
+[MASTER]
+ignore=CVS
+good-names=logger,e,i,j,n,m,f,_
+
+[MESSAGES CONTROL]
+disable=all
+enable=unused-import,
+       fixme,
+       useless-object-inheritance,
+       unused-variable,
+       unused-argument,
+       unexpected-keyword-arg,
+       string,
+       unreachable,
+       invalid-name,
+       logging-not-lazy,
+       unnecesary-pass
