+++ setup.py
@@ -1,47 +1,15 @@
-# -*- coding: utf-8 -*-
+from sys import exit
+from setuptools import setup, version
 
 
-from setuptools import (
-    find_packages,
-    setup,
-)
-
+if version.__version__ < "40.1.0":
+    exit(
+        "At least version 40.1.0 of setuptools is required ({} is installed)".format(
+            version.__version__
+        )
+    )
 
 setup(
-    name='pytest-docker',
-    url='https://github.com/AndreLouisCaron/pytest-docker',
-    version='0.7.0',
-    license='MIT',
-    maintainer='Andre Caron',
-    maintainer_email='andre.l.caron@gmail.com',
-    classifiers=[
-        'Framework :: Pytest',
-        'Development Status :: 3 - Alpha',
-        'License :: OSI Approved :: MIT License',
-        'Programming Language :: Python',
-        'Programming Language :: Python :: 2.7',
-        'Programming Language :: Python :: 3.5',
-        'Topic :: Utilities',
-        'Intended Audience :: Developers',
-        'Operating System :: Unix',
-        'Operating System :: POSIX',
-        'Operating System :: Microsoft :: Windows',
-    ],
-    keywords=[
-        'docker',
-        'docker-compose',
-        'pytest',
-    ],
-    packages=find_packages(where='src'),
-    package_dir={
-        '': 'src',
-    },
-    entry_points={
-        'pytest11': [
-            'docker = pytest_docker',
-        ],
-    },
-    install_requires=[
-        'attrs>=16',
-    ],
+    setup_requires=["wheel >= 0.32", "pytest-runner >= 5.0, <6.0"],
+    entry_points={"pytest11": ["docker = pytest_docker"]},
 )
