+++ tests/test_fixtures.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 import os.path
 
 
@@ -8,3 +6,7 @@ HERE = os.path.dirname(os.path.abspath(__file__))
 
 def test_docker_compose_file(docker_compose_file):
     assert docker_compose_file == os.path.join(HERE, "docker-compose.yml")
+
+
+def test_docker_compose_project(docker_compose_project_name):
+    assert docker_compose_project_name == "pytest{}".format(os.getpid())
